Fix clang compilation by removing "register" keywords
The clang compiler warns about the "register" keyword being deprecated in C++11. As the Jellyfish build system treats warnings as errors, this breaks the build with clang. I suggest to remove the keyword instead of silencing the warning because most modern compilers (e.g. the recent versions of GCC, clang and Visual Studio) ignore the keyword in their optimization algorithms anyway.